Maybe fix macro chain crash
We've got a report of a macro chain crash issue. I haven't been able to
reproduce it but from the stack traces I've seen it seems to be breaking
when copying `IsEmpty` from the macro to the RaptureShellModule macro
line buffer.

From that I'm assuming there's an issue with the data we're passing to
`ReplaceMacroLines` which might be causing invalid memory. We know that
`SetText`/`SetString` without a null terminator was causing issues
elsewhere so it seems reasonable that it might be causing issues here.

If this doesn't fix the issue then we're going to just keep digging.